# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition"
  },
...
  "paths" : {
...
    "/v1/provenance" : {
      "post" : {
        "tags" : [ "provenance-controller" ],
        "operationId" : "record",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProvenanceRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/v1/provenance/bulk" : {
      "post" : {
        "tags" : [ "provenance-controller" ],
        "operationId" : "recordBatch",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BulkProvenanceRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BatchResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/consent" : {
      "post" : {
        "tags" : [ "consent-controller" ],
        "operationId" : "capture",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CaptureRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CaptureResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/provenance/{id}/substantiate" : {
      "post" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "substantiate",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SubstantiateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Record"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/rights/summary" : {
      "get" : {
        "tags" : [ "rights-controller" ],
        "operationId" : "summary",
        "parameters" : [ {
          "name" : "entityId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TypeSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/rights/subject/{entityId}/{subjectId}" : {
      "get" : {
        "tags" : [ "rights-controller" ],
        "operationId" : "forSubject",
        "parameters" : [ {
          "name" : "entityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "subjectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Request"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/receipts" : {
      "get" : {
        "tags" : [ "receipt-controller" ],
        "operationId" : "forSubject_1",
        "parameters" : [ {
          "name" : "entityId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "subjectId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/StoredReceipt"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/provenance/{entityId}/{subjectId}" : {
      "get" : {
        "tags" : [ "provenance-controller" ],
        "operationId" : "forSubject_2",
        "parameters" : [ {
          "name" : "entityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "subjectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Record"
                }
              }
            }
          }
        }
      }
    },
    "/v1/provenance/summary" : {
      "get" : {
        "tags" : [ "provenance-controller" ],
        "operationId" : "summary_1",
        "parameters" : [ {
          "name" : "entityId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SourceSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/provenance/summary" : {
      "get" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "provenanceSummary",
        "parameters" : [ {
          "name" : "entityId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SourceSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/provenance/quarantined" : {
      "get" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "quarantined",
        "parameters" : [ {
          "name" : "entityId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 100
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Record"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Request" : {
        "type" : "object",
        "properties" : {
          "requestId" : {
            "type" : "string"
          },
          "entityId" : {
            "type" : "string"
          },
          "subjectId" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "ACCESS", "CORRECTION", "COMPLETION", "ERASURE", "NOMINATION", "GRIEVANCE", "CONSENT_WITHDRAWAL", "PORTABILITY", "OPT_OUT_OF_SALE" ]
          },
          "jurisdiction" : {
            "type" : "string",
            "enum" : [ "IN", "EU", "UK", "SG", "MY", "KR", "US_CA", "US_CO", "US_CT", "US_TX", "US_OR", "US_MT", "US_DE", "US_NJ", "US_NE", "US_NH", "US_MN", "US_MD", "US_VA", "US_UT", "US_IA", "OTHER" ]
          },
          "status" : {
            "type" : "string",
            "enum" : [ "RECEIVED", "IN_PROGRESS", "AWAITING_SUBJECT", "FULFILLED", "REJECTED", "WITHDRAWN" ]
          },
          "receivedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dueAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dueAtBasis" : {
            "type" : "string"
          },
          "closedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "acknowledgedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "assignedTo" : {
            "type" : "string"
          },
          "resolution" : {
            "type" : "string"
          },
          "details" : {
            "type" : "string"
          },
          "verification" : {
            "type" : "string",
            "enum" : [ "PORTAL_TOKEN", "OPERATOR_ASSERTED", "UNVERIFIED" ]
          },
          "verifiedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "verificationDetail" : {
            "type" : "string"
          }
        }
      },
...
      "ProvenanceRequest" : {
        "type" : "object",
        "properties" : {
          "entityId" : {
            "type" : "string"
          },
          "subjectId" : {
            "type" : "string"
          },
          "identifierType" : {
            "type" : "string",
            "enum" : [ "PHONE", "EMAIL", "EMPLOYEE_ID", "CANDIDATE_ID", "DEVICE_ID", "EXTERNAL_ID" ]
          },
          "identifierValue" : {
            "type" : "string"
          },
          "sourceType" : {
            "type" : "string",
            "enum" : [ "DIRECT_COLLECTION", "CLIENT_SUPPLIED", "REFERRAL", "EVENT_OR_TRADESHOW", "PURCHASED_LIST", "APPENDED", "PUBLIC_SOURCE", "WEB_SCRAPED", "LEGACY_UNKNOWN" ]
          },
          "sourceName" : {
            "type" : "string",
            "minLength" : 1
          },
          "acquiredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "originalLegalBasis" : {
            "type" : "string",
            "enum" : [ "CONSENT", "INFERRED_CONSENT", "LEGITIMATE_USE_EMPLOYMENT", "LEGITIMATE_USE_VOLUNTARY", "CONTRACT_PERFORMANCE", "LEGAL_OBLIGATION", "LEGITIMATE_INTEREST", "VITAL_INTEREST", "PUBLIC_INTEREST" ]
          },
          "evidenceRef" : {
            "type" : "string"
          },
          "contractRef" : {
            "type" : "string"
          }
        },
        "required" : [ "acquiredAt", "sourceName", "sourceType" ]
      },
      "Result" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "subjectId" : {
            "type" : "string"
          },
          "inserted" : {
            "type" : "boolean"
          },
          "accepted" : {
            "type" : "boolean"
          },
          "rejectionReason" : {
            "type" : "string"
          }
        }
      },
      "BulkProvenanceRequest" : {
        "type" : "object",
        "properties" : {
          "entityId" : {
            "type" : "string",
            "minLength" : 1
          },
          "batchRef" : {
            "type" : "string"
          },
          "records" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ProvenanceRequest"
            },
            "maxItems" : 5000,
            "minItems" : 0
          }
        },
        "required" : [ "entityId", "records" ]
      },
      "BatchResult" : {
        "type" : "object",
        "properties" : {
          "submitted" : {
            "type" : "integer",
            "format" : "int32"
          },
          "inserted" : {
            "type" : "integer",
            "format" : "int32"
          },
          "duplicates" : {
            "type" : "integer",
            "format" : "int32"
          },
          "rejected" : {
            "type" : "integer",
            "format" : "int32"
          },
          "results" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Result"
            }
          }
        }
      },
...
      "CaptureRequest" : {
        "type" : "object",
        "properties" : {
          "entityId" : {
            "type" : "string",
            "minLength" : 1
          },
          "subjectId" : {
            "type" : "string"
          },
          "subject" : {
            "$ref" : "#/components/schemas/SubjectRef"
          },
          "alsoKnownAs" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SubjectRef"
            }
          },
          "jurisdiction" : {
            "type" : "string",
            "enum" : [ "IN", "EU", "UK", "SG", "MY", "KR", "US_CA", "US_CO", "US_CT", "US_TX", "US_OR", "US_MT", "US_DE", "US_NJ", "US_NE", "US_NH", "US_MN", "US_MD", "US_VA", "US_UT", "US_IA", "OTHER" ]
          },
          "languageTag" : {
            "type" : "string",
            "minLength" : 1
          },
          "channel" : {
            "type" : "string",
            "enum" : [ "WEB", "MOBILE_APP", "VOICE_CALL", "SMS", "WHATSAPP", "EMAIL", "KIOSK", "PAPER", "API", "BULK_IMPORT", "POSTAL" ]
          },
          "applicationId" : {
            "type" : "string"
          },
          "captureMethod" : {
            "type" : "string",
            "enum" : [ "CHECKBOX_OPT_IN", "CLICK_THROUGH", "DOUBLE_OPT_IN", "OTP_VERIFIED", "VERBAL_RECORDED", "WET_SIGNATURE", "BIOMETRIC_ENROLMENT", "PARENTAL_VERIFIED", "RELAYED_BY_CONSENT_MANAGER", "IMPORTED_WITH_PROVENANCE", "INFERRED_FROM_RELATIONSHIP", "NOT_APPLICABLE" ]
          },
          "actorType" : {
            "type" : "string",
            "enum" : [ "SUBJECT", "PARENT_GUARDIAN", "ADMIN", "AGENT", "SYSTEM", "IMPORT", "REGULATOR", "CONSENT_MANAGER" ]
          },
          "actorId" : {
            "type" : "string"
          },
          "noticeId" : {
            "type" : "string"
          },
          "noticeVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "choices" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PurposeChoiceDto"
            },
            "minItems" : 1
          },
          "rejectAllOffered" : {
            "type" : "boolean"
          },
          "occurredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "idempotencyKey" : {
            "type" : "string"
          },
          "evidenceRef" : {
            "type" : "string"
          },
          "attributes" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "guardianVerification" : {
            "$ref" : "#/components/schemas/GuardianVerificationDto"
          }
        },
        "required" : [ "actorType", "captureMethod", "choices", "entityId", "jurisdiction", "languageTag" ]
      },
      "GuardianVerificationDto" : {
        "type" : "object",
        "properties" : {
          "method" : {
            "type" : "string",
            "enum" : [ "EXISTING_VERIFIED_ACCOUNT", "DIGILOCKER_VIRTUAL_TOKEN", "OTHER_DOCUMENTED" ]
          },
          "reference" : {
            "type" : "string",
            "minLength" : 1
          },
          "verifiedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "verifiedBy" : {
            "type" : "string"
          }
        },
        "required" : [ "method", "reference", "verifiedAt" ]
      },
      "PurposeChoiceDto" : {
        "type" : "object",
        "properties" : {
          "purposeCode" : {
            "type" : "string",
            "minLength" : 1
          },
          "granted" : {
            "type" : "boolean"
          },
          "preTicked" : {
            "type" : "boolean"
          },
          "separateAction" : {
            "type" : "boolean"
          }
        },
        "required" : [ "purposeCode" ]
      },
      "SubjectRef" : {
        "type" : "object",
        "properties" : {
          "identifierType" : {
            "type" : "string",
            "enum" : [ "PHONE", "EMAIL", "EMPLOYEE_ID", "CANDIDATE_ID", "DEVICE_ID", "EXTERNAL_ID" ]
          },
          "value" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "identifierType", "value" ]
      },
      "CaptureResponse" : {
        "type" : "object",
        "properties" : {
          "accepted" : {
            "type" : "boolean"
          },
          "subjectId" : {
            "type" : "string"
          },
          "events" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EventSummary"
            }
          },
          "violations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ViolationDto"
            }
          }
        }
      },
      "EventSummary" : {
        "type" : "object",
        "properties" : {
          "eventId" : {
            "type" : "string"
          },
          "purposeCode" : {
            "type" : "string"
          },
          "purposeVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "eventType" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "GRANTED", "DENIED", "WITHDRAWN", "NOT_ASKED", "EXPIRED", "INVALIDATED", "PENDING_SYNC", "CONFLICTED", "UNKNOWN" ]
          },
          "occurredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "sequenceNumber" : {
            "type" : "integer",
            "format" : "int64"
          },
          "eventHash" : {
            "type" : "string"
          }
        }
      },
      "ViolationDto" : {
        "type" : "object",
        "properties" : {
          "purposeCode" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          },
          "detail" : {
            "type" : "string"
          }
        }
      },
...
      "SubstantiateRequest" : {
        "type" : "object",
        "properties" : {
          "evidenceNote" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "evidenceNote" ]
      },
      "Record" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "entityId" : {
            "type" : "string"
          },
          "subjectId" : {
            "type" : "string"
          },
          "sourceType" : {
            "type" : "string"
          },
          "sourceName" : {
            "type" : "string"
          },
          "acquiredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "originalLegalBasis" : {
            "type" : "string"
          },
          "contractRef" : {
            "type" : "string"
          },
          "substantiated" : {
            "type" : "boolean"
          },
          "quarantined" : {
            "type" : "boolean"
          }
        }
      },
...
      "TypeSummary" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "enum" : [ "ACCESS", "CORRECTION", "COMPLETION", "ERASURE", "NOMINATION", "GRIEVANCE", "CONSENT_WITHDRAWAL", "PORTABILITY", "OPT_OUT_OF_SALE" ]
          },
          "open" : {
            "type" : "integer",
            "format" : "int64"
          },
          "overdue" : {
            "type" : "integer",
            "format" : "int64"
          },
          "closed" : {
            "type" : "integer",
            "format" : "int64"
          },
          "closedLate" : {
            "type" : "integer",
            "format" : "int64"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "StoredReceipt" : {
        "type" : "object",
        "properties" : {
          "receiptId" : {
            "type" : "string"
          },
          "entityId" : {
            "type" : "string"
          },
          "subjectId" : {
            "type" : "string"
          },
          "issuedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "payload" : {
            "type" : "string"
          },
          "payloadHash" : {
            "type" : "string"
          },
          "evidenceHash" : {
            "type" : "string"
          },
          "noticeId" : {
            "type" : "string"
          },
          "noticeVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "languageTag" : {
            "type" : "string"
          },
          "purposeCount" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "SourceSummary" : {
        "type" : "object",
        "properties" : {
          "sourceType" : {
            "type" : "string"
          },
          "sourceName" : {
            "type" : "string"
          },
          "quarantined" : {
            "type" : "integer",
            "format" : "int64"
          },
          "contactable" : {
            "type" : "integer",
            "format" : "int64"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
    }
  }
}